randomQuote refactoring
