+++ server/src/main/java/todo/todominsung/module/service/RandomQuoteService.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.annotation.PostConstruct;
 import org.springframework.stereotype.Service;
-import todo.todominsung.module.entity.RandomQuote;
+import todo.todominsung.module.dto.RandomQuoteDto;
 
 import java.io.InputStream;
 import java.util.*;
@@ -12,20 +12,20 @@ import java.util.*;
 @Service
 public class RandomQuoteService {
 
-    private List<RandomQuote> quotes;
+    private List<RandomQuoteDto> quotes;
 
     @PostConstruct // 빈이 초기화된 후 자동으로 실행
     public void loadQuotes() {
         try {
             ObjectMapper objectMapper = new ObjectMapper(); // JSON -> Java 객체 , Java 객체 -> JSON으로 변환할 때 사용
             InputStream inputStream = getClass().getResourceAsStream("/quotes.json");
-            quotes = objectMapper.readValue(inputStream, new TypeReference<List<RandomQuote>>() {});
+            quotes = objectMapper.readValue(inputStream, new TypeReference<List<RandomQuoteDto>>() {});
         } catch (Exception error) {
             throw new RuntimeException("Failed to load quotes.json", error);
         }
     }
 
-    public RandomQuote getRandomQuote() {
+    public RandomQuoteDto getRandomQuote() {
         int randomIndex = (int) (Math.random() * quotes.size());
         return quotes.get(randomIndex);
     }
